telegram-bot: update NewDentalWork need to complete

+++ telegram-bot/src/main/java/org/lab/telegram_bot/domain/command/impl/ProductMapHandler.java
@@ -20,7 +20,6 @@ import org.telegram.telegrambots.meta.api.objects.CallbackQuery;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
@@ -53,12 +52,12 @@ public class ProductMapHandler extends BotCommandHandler {
         String messageText = message.getText();
         Steps step = getStep(session);
         return switch (step) {
-            case GET -> get(session, locale);
-            case UPDATE_LIST -> updateList(session, locale, message.getMessageId());
-            case SELECT_TO_UPDATE -> selectToUpdate(session, locale, messageText, message.getMessageId());
-            case UPDATE_INPUT -> updateInput(session, locale, messageText);
-            case DELETE_LIST -> deleteList(session, locale, message.getMessageId());
-            case SELECT_TO_DELETE -> selectToDelete(session, locale, messageText, message.getMessageId());
+            case GET_PRODUCT_TYPES -> get(session, locale);
+            case PRODUCT_TYPE_LIST_FOR_UPDATE -> updateList(session, locale, message.getMessageId());
+            case SELECT_PRODUCT_TYPE_TO_UPDATE -> selectToUpdate(session, locale, messageText, message.getMessageId());
+            case UPDATE_PRODUCT_TYPE_INPUT -> updateInput(session, locale, messageText);
+            case PRODUCT_TYPE_LIST_FOR_DELETE -> deleteList(session, locale, message.getMessageId());
+            case SELECT_PRODUCT_TYPE_TO_DELETE -> selectToDelete(session, locale, messageText, message.getMessageId());
         };
     }
 
@@ -67,28 +66,32 @@ public class ProductMapHandler extends BotCommandHandler {
         String messageText = callbackQuery.getData();
         Steps step = getStep(session);
         return switch (step) {
-            case GET -> get(session, locale);
-            case UPDATE_LIST -> updateList(session, locale, callbackQuery.getMessage().getMessageId());
-            case SELECT_TO_UPDATE ->
+            case GET_PRODUCT_TYPES -> get(session, locale);
+            case PRODUCT_TYPE_LIST_FOR_UPDATE -> updateList(session, locale, callbackQuery.getMessage().getMessageId());
+            case SELECT_PRODUCT_TYPE_TO_UPDATE ->
                     selectToUpdate(session, locale, messageText, callbackQuery.getMessage().getMessageId());
-            case UPDATE_INPUT -> updateInput(session, locale, messageText);
-            case DELETE_LIST -> deleteList(session, locale, callbackQuery.getMessage().getMessageId());
-            case SELECT_TO_DELETE ->
+            case UPDATE_PRODUCT_TYPE_INPUT -> updateInput(session, locale, messageText);
+            case PRODUCT_TYPE_LIST_FOR_DELETE -> deleteList(session, locale, callbackQuery.getMessage().getMessageId());
+            case SELECT_PRODUCT_TYPE_TO_DELETE ->
                     selectToDelete(session, locale, messageText, callbackQuery.getMessage().getMessageId());
         };
     }
 
-    //TODO back query process
 
     private SendMessage get(ChatSession session, Locale locale) {
         ProductMap productMap = productMapService.findAll(session.getUserId());
         return sendProductMapAsMessage(productMap, session, locale);
     }
 
+    private EditMessageText get(ChatSession session, Locale locale, int messageId) {
+        ProductMap productMap = productMapService.findAll(session.getUserId());
+        return sendProductMapAsMessage(productMap, session, locale, messageId);
+    }
+
     private EditMessageText updateList(ChatSession session, Locale locale, int messageId) {
-        String text = messageSource.getMessage(Steps.UPDATE_LIST.name(), null, locale);
-        InlineKeyboardMarkup keyboardMarkup = productMapAsCallbackQuery(session, locale, Steps.SELECT_TO_UPDATE);
-        session.setStep(Steps.SELECT_TO_UPDATE.ordinal());
+        String text = messageSource.getMessage(Steps.PRODUCT_TYPE_LIST_FOR_UPDATE.name(), null, locale);
+        InlineKeyboardMarkup keyboardMarkup = productMapAsCallbackQuery(session, locale, Steps.SELECT_PRODUCT_TYPE_TO_UPDATE);
+        session.setStep(Steps.SELECT_PRODUCT_TYPE_TO_UPDATE.ordinal());
         session.setCommand(BotCommands.PRODUCT_MAP);
         chatSessionService.save(session);
         return editMessageText(session.getChatId(), messageId, text, keyboardMarkup);
@@ -96,11 +99,17 @@ public class ProductMapHandler extends BotCommandHandler {
 
     private EditMessageText selectToUpdate(ChatSession session, Locale locale, String messageText, int messageId) {
         String[] callbackData = ChatBotUtility.callBackQueryParse(messageText);
+        if (callbackData[2].equals(ButtonKeys.CANCEL.name())) {
+            String text = messageSource.getMessage(CANCEL_RESPONSE, null, locale);
+            session.clear();
+            chatSessionService.save(session);
+            return editMessageText(session.getChatId(), messageId, text);
+        }
         UUID productTypeId = UUID.fromString(callbackData[2]);
         ProductType productType = productMapService.findById(productTypeId, session.getUserId());
-        session.addAttribute(Attributes.PRODUCT_TYPE_TO_UPDATE.name(), productTypeId.toString());
-        String text = messageSource.getMessage(Steps.SELECT_TO_UPDATE.name(), new Object[]{productType.getTitle()}, locale);
-        session.setStep(Steps.UPDATE_INPUT.ordinal());
+        session.addAttribute(Attributes.PRODUCT_TYPE_ID, productTypeId);
+        String text = messageSource.getMessage(Steps.SELECT_PRODUCT_TYPE_TO_UPDATE.name(), new Object[]{productType.getTitle()}, locale);
+        session.setStep(Steps.UPDATE_PRODUCT_TYPE_INPUT.ordinal());
         session.setCommand(BotCommands.PRODUCT_MAP);
         chatSessionService.save(session);
         return editMessageText(session.getChatId(), messageId, text);
@@ -108,19 +117,19 @@ public class ProductMapHandler extends BotCommandHandler {
 
     private SendMessage updateInput(ChatSession session, Locale locale, String messageText) {
         float newPrice = Float.parseFloat(messageText);
-        String productMapId = session.getAttribute(Attributes.PRODUCT_TYPE_TO_UPDATE.name());
-        if (productMapId == null || productMapId.isBlank()) {
-            throw new IllegalArgumentException("ChatSession attribute '%s' is null or empty".formatted(Attributes.PRODUCT_TYPE_TO_UPDATE));
+        UUID productTypeId = (UUID) session.getAttribute(Attributes.PRODUCT_TYPE_ID);
+        if (productTypeId == null) {
+            throw new IllegalArgumentException("ChatSession attribute '%s' is null".formatted(Attributes.PRODUCT_TYPE_ID));
         }
-        UUID id = UUID.fromString(productMapId);
-        ProductMap productMap = productMapService.updatePrice(id, newPrice, session.getUserId());
+        ProductMap productMap = productMapService.updatePrice(productTypeId, newPrice, session.getUserId());
+        session.removeAttribute(Attributes.PRODUCT_TYPE_ID);
         return sendProductMapAsMessage(productMap, session, locale);
     }
 
     private EditMessageText deleteList(ChatSession session, Locale locale, int messageId) {
-        String text = messageSource.getMessage(Steps.DELETE_LIST.name(), null, locale);
-        InlineKeyboardMarkup keyboardMarkup = productMapAsCallbackQuery(session, locale, Steps.SELECT_TO_DELETE);
-        session.setStep(Steps.SELECT_TO_DELETE.ordinal());
+        String text = messageSource.getMessage(Steps.PRODUCT_TYPE_LIST_FOR_DELETE.name(), null, locale);
+        InlineKeyboardMarkup keyboardMarkup = productMapAsCallbackQuery(session, locale, Steps.SELECT_PRODUCT_TYPE_TO_DELETE);
+        session.setStep(Steps.SELECT_PRODUCT_TYPE_TO_DELETE.ordinal());
         session.setCommand(BotCommands.PRODUCT_MAP);
         chatSessionService.save(session);
         return editMessageText(session.getChatId(), messageId, text, keyboardMarkup);
@@ -128,6 +137,9 @@ public class ProductMapHandler extends BotCommandHandler {
 
     private EditMessageText selectToDelete(ChatSession session, Locale locale, String messageText, int messageId) {
         String[] callbackData = ChatBotUtility.callBackQueryParse(messageText);
+        if (callbackData[2].equals(ButtonKeys.BACK.name())) {
+            return get(session, locale, messageId);
+        }
         UUID productTypeId = UUID.fromString(callbackData[2]);
         ProductMap productMap = productMapService.delete(productTypeId, session.getUserId());
         return sendProductMapAsMessage(productMap, session, locale, messageId);
@@ -135,9 +147,9 @@ public class ProductMapHandler extends BotCommandHandler {
 
     private SendMessage sendProductMapAsMessage(ProductMap productMap, ChatSession session, Locale locale) {
         String text = productMapAsMessageText(productMap);
-        String callbackQueryPrefix = ChatBotUtility.callBackQueryPrefix(BotCommands.PRODUCT_MAP, Steps.UPDATE_LIST.ordinal());
+        String callbackQueryPrefix = ChatBotUtility.callBackQueryPrefix(BotCommands.PRODUCT_MAP, Steps.PRODUCT_TYPE_LIST_FOR_UPDATE.ordinal());
         InlineKeyboardButton update = keyboardBuilderKit.callbackButton(ButtonKeys.UPDATE, callbackQueryPrefix, locale);
-        callbackQueryPrefix = ChatBotUtility.callBackQueryPrefix(BotCommands.PRODUCT_MAP, Steps.DELETE_LIST.ordinal());
+        callbackQueryPrefix = ChatBotUtility.callBackQueryPrefix(BotCommands.PRODUCT_MAP, Steps.PRODUCT_TYPE_LIST_FOR_DELETE.ordinal());
         InlineKeyboardButton delete = keyboardBuilderKit.callbackButton(ButtonKeys.DELETE, callbackQueryPrefix, locale);
         var keyboard = keyboardBuilderKit.inlineKeyboard(List.of(update, delete));
         session.setStep(0);
@@ -148,9 +160,9 @@ public class ProductMapHandler extends BotCommandHandler {
 
     private EditMessageText sendProductMapAsMessage(ProductMap productMap, ChatSession session, Locale locale, int messageId) {
         String text = productMapAsMessageText(productMap);
-        String callbackQueryPrefix = ChatBotUtility.callBackQueryPrefix(BotCommands.PRODUCT_MAP, Steps.UPDATE_LIST.ordinal());
+        String callbackQueryPrefix = ChatBotUtility.callBackQueryPrefix(BotCommands.PRODUCT_MAP, Steps.PRODUCT_TYPE_LIST_FOR_UPDATE.ordinal());
         InlineKeyboardButton update = keyboardBuilderKit.callbackButton(ButtonKeys.UPDATE, callbackQueryPrefix, locale);
-        callbackQueryPrefix = ChatBotUtility.callBackQueryPrefix(BotCommands.PRODUCT_MAP, Steps.DELETE_LIST.ordinal());
+        callbackQueryPrefix = ChatBotUtility.callBackQueryPrefix(BotCommands.PRODUCT_MAP, Steps.PRODUCT_TYPE_LIST_FOR_DELETE.ordinal());
         InlineKeyboardButton delete = keyboardBuilderKit.callbackButton(ButtonKeys.DELETE, callbackQueryPrefix, locale);
         var keyboard = keyboardBuilderKit.inlineKeyboard(List.of(update, delete));
         session.setStep(0);
@@ -169,8 +181,8 @@ public class ProductMapHandler extends BotCommandHandler {
             buttons.add(new ArrayList<>());
             buttons.getLast().add(button);
         }
-        InlineKeyboardButton backButton = keyboardBuilderKit.callbackButton(ButtonKeys.BACK, callbackQueryPrefix, locale);
-        buttons.add(List.of(backButton));
+        InlineKeyboardButton cancelButton = keyboardBuilderKit.callbackButton(ButtonKeys.CANCEL, callbackQueryPrefix, locale);
+        buttons.add(List.of(cancelButton));
         return keyboardBuilderKit.inlineKeyboard(buttons);
     }
 
@@ -195,15 +207,15 @@ public class ProductMapHandler extends BotCommandHandler {
 
 
     enum Steps {
-        GET,
-        UPDATE_LIST,
-        SELECT_TO_UPDATE,
-        UPDATE_INPUT,
-        DELETE_LIST,
-        SELECT_TO_DELETE
+        GET_PRODUCT_TYPES,
+        PRODUCT_TYPE_LIST_FOR_UPDATE,
+        SELECT_PRODUCT_TYPE_TO_UPDATE,
+        UPDATE_PRODUCT_TYPE_INPUT,
+        PRODUCT_TYPE_LIST_FOR_DELETE,
+        SELECT_PRODUCT_TYPE_TO_DELETE
     }
 
     enum Attributes {
-        PRODUCT_TYPE_TO_UPDATE
+        PRODUCT_TYPE_ID
     }
 }
